+++ zenref/src/Program.cs
@@ -9,42 +9,6 @@ namespace P3Project
         }
     }
 
-    public class Filter
-    {
-        public string createFilter()
-        {
-            //UI? Button testing?
-            throw new NotImplementedException();
-        }
-
-        public string saveFilter()
-        {
-            throw new NotImplementedException();
-        }
-
-        public string editFilter()
-        {
-            throw new NotImplementedException();
-        }
-
-        public string deleteFilter()
-        {
-            throw new NotImplementedException();
-        }
-        
-        public string searchFilter()
-        {
-            //something something, leder efter en bestemt streng eller et bestemt felt
-            throw new NotImplementedException();
-        }
-
-        public string categorize()
-        {
-            //something something, tager searching metoden og på hits sætter den referencen over i den nye fane
-            throw new NotImplementedException();
-        }
-    }
-
     public class Spreadsheet
     {
